Factorizing code
