+++ core/IO.cpp
@@ -1,0 +1,33 @@
+#include "IO.h"
+
+IO::IO(QIODevice *device) :
+	_device(device)
+{
+}
+
+IO::~IO()
+{
+}
+
+bool IO::canRead() const
+{
+	if(_device) {
+		if(!_device->isOpen()) {
+			return _device->open(QIODevice::ReadOnly);
+		}
+		return _device->isReadable();
+	}
+	return false;
+}
+
+bool IO::canWrite() const
+{
+	if(_device) {
+		if(!_device->isOpen()) {
+			return _device->open(QIODevice::WriteOnly);
+		}
+		return _device->isWritable();
+	}
+	return false;
+}
+
