+++ core/field/BackgroundTilesIO.h
@@ -19,6 +19,7 @@
 #define BACKGROUNDTILESIO_H
 
 #include <QtCore>
+#include "../IO.h"
 #include "BackgroundTiles.h"
 
 #define MAX_TILE_DST	1024
@@ -77,30 +78,17 @@ typedef struct {
 	quint8 state;
 } paramTile;
 
-class BackgroundTilesIO
+class BackgroundTilesIO : public IO
 {
 public:
 	explicit BackgroundTilesIO(QIODevice *device);
-	virtual ~BackgroundTilesIO();
-
-	inline void setDevice(QIODevice *device) {
-		_device = device;
-	}
-
-	inline QIODevice *device() const {
-		return _device;
-	}
-
-	bool canRead() const;
-	bool canWrite() const;
+	virtual ~BackgroundTilesIO() {}
 
 	bool read(BackgroundTiles &tiles) const;
 	bool write(const BackgroundTiles &tiles) const;
 protected:
 	virtual bool readData(BackgroundTiles &tiles) const=0;
 	virtual bool writeData(const BackgroundTiles &tiles) const=0;
-private:
-	QIODevice *_device;
 };
 
 class BackgroundTilesIOPC : public BackgroundTilesIO
